Change UbiBot uplink values to float

+++ src/main/java/net/kf03w5t5741l/sensorbase/server/domain/UbiBotUplink.java
@@ -15,16 +15,16 @@ public class UbiBotUplink {
     private int channelId;
     private LocalDateTime CreatedAt;
 
-    private int field1;
-    private int field2;
-    private int field3;
-    private int field4;
-    private int field5;
-    private int field6;
-    private int field7;
-    private int field8;
-    private int field9;
-    private int field10;
+    private float field1;
+    private float field2;
+    private float field3;
+    private float field4;
+    private float field5;
+    private float field6;
+    private float field7;
+    private float field8;
+    private float field9;
+    private float field10;
     private String status;
 
     public Long getUbibotUplinkId() {
@@ -51,83 +51,83 @@ public class UbiBotUplink {
         CreatedAt = createdAt;
     }
 
-    public int getField1() {
+    public float getField1() {
         return field1;
     }
 
-    public void setField1(int field1) {
+    public void setField1(float field1) {
         this.field1 = field1;
     }
 
-    public int getField2() {
+    public float getField2() {
         return field2;
     }
 
-    public void setField2(int field2) {
+    public void setField2(float field2) {
         this.field2 = field2;
     }
 
-    public int getField3() {
+    public float getField3() {
         return field3;
     }
 
-    public void setField3(int field3) {
+    public void setField3(float field3) {
         this.field3 = field3;
     }
 
-    public int getField4() {
+    public float getField4() {
         return field4;
     }
 
-    public void setField4(int field4) {
+    public void setField4(float field4) {
         this.field4 = field4;
     }
 
-    public int getField5() {
+    public float getField5() {
         return field5;
     }
 
-    public void setField5(int field5) {
+    public void setField5(float field5) {
         this.field5 = field5;
     }
 
-    public int getField6() {
+    public float getField6() {
         return field6;
     }
 
-    public void setField6(int field6) {
+    public void setField6(float field6) {
         this.field6 = field6;
     }
 
-    public int getField7() {
+    public float getField7() {
         return field7;
     }
 
-    public void setField7(int field7) {
+    public void setField7(float field7) {
         this.field7 = field7;
     }
 
-    public int getField8() {
+    public float getField8() {
         return field8;
     }
 
-    public void setField8(int field8) {
+    public void setField8(float field8) {
         this.field8 = field8;
     }
 
-    public int getField9() {
+    public float getField9() {
         return field9;
     }
 
-    public void setField9(int field9) {
+    public void setField9(float field9) {
         this.field9 = field9;
     }
 
-    public int getField10() {
+    public float getField10() {
         return field10;
     }
 
-    public void setField10(int field10) {
+    public void setField10(float field10) {
         this.field10 = field10;
     }
